requirements.txt
pyqt5
matplotlib
pyserial
bleak
pandas
numpy
scipy
QDarkStyle
QtAwesome